Added the table for the Active Time, and all of the code to use it.  However this casued an error.
Caused by: java.lang.IllegalStateException: A migration from 1 to 2 was required but not found. Please provide the necessary Migration path via RoomDatabase.Builder.addMigration(Migration ...) or allow for destructive migrations via one of the RoomDatabase.Builder.fallbackToDestructiveMigration* methods.

+++ app/src/main/java/com/example/fitnessapp/StepCountDao.java
@@ -22,7 +22,7 @@ public interface StepCountDao {
     LiveData<List<StepCount>> getAll();
 
     @Query("SELECT totalsteps FROM stepcount WHERE date = :date")
-    public abstract long currentCount(String date);
+    public abstract int currentCount(String date);
 
     @Query("SELECT * FROM stepcount where date > :date")
     public abstract List<StepCount> allFromDate(String date);
